+++ components/SheetMusicApp/SheetMusicApp.js
@@ -34,23 +34,23 @@ const SheetMusicApp = ({ slug, className }) => {
   return (
     <App className={[css.sheetMusicApp, className || ''].join(' ')}>
       <header className={css.header}>
-        <DXLabLogo className={css.dxlabLogo} />
+        {/* <DXLabLogo className={css.dxlabLogo} /> */}
 
-        <span className={css.headerDivider}></span>
+        {/* <span className={css.headerDivider}></span> */}
 
         <h1 className={css.virtuosoLogo}>
           Virtu<strong>OSO</strong>
-          <svg width="2em" height="1.5em">
+          <svg width="2em" height="1.4em">
             <ellipse
               cx="0.15em"
-              cy="0.9em"
+              cy="0.8em"
               rx="0.2em"
               ry="0.3em"
               style={{ fill: 'var(--colour-primary)' }}
             />
             <ellipse
               cx="1.5em"
-              cy="0.9em"
+              cy="0.8em"
               rx="0.2em"
               ry="0.3em"
               style={{ fill: 'var(--colour-primary)' }}
@@ -73,13 +73,26 @@ const SheetMusicApp = ({ slug, className }) => {
               height={3}
               style={{ fill: 'var(--colour-primary)' }}
             />
+            <rect
+              x="0.3em"
+              y={5}
+              width="1.4em"
+              height={1}
+              style={{ fill: 'var(--colour-primary)' }}
+            />
           </svg>
         </h1>
 
         <p className={css.virtuosoLogoSub}>
-          <strong>O</strong>nline <strong>S</strong>heet <strong>O</strong>
+          <strong>O</strong>nline
+          <br />
+          <strong>S</strong>heet
+          <br />
+          <strong>O</strong>
           rchestra
         </p>
+
+        <span className={css.headerDivider}></span>
 
         {/* <p>♩ ♪ ♫ ♬ ♭</p> */}
 
@@ -109,6 +122,8 @@ const SheetMusicApp = ({ slug, className }) => {
             })}
           />
         </Sidebar>
+
+        <DXLabLogo className={css.dxlabLogo} />
       </header>
 
       <div className={css.content}>